Clean up approach to handle gallery.
Replaced methods to handle previous, next and change features.
Make easier to read, improve single responsibility.

Complication comes to passing change photo to thumbnails. Can't be don't
the current way. Might need to have another method as a closure.

Not sure I like it. Maybe easier to read but is more code that needs to
be shared with the whole class. Less pure than before.